feat: rss endpoint
