+++ src/markdoc/layouts/Category.svelte
@@ -1,17 +1,8 @@
-<script context="module" lang="ts">
-	export type CategoryData = {
-		name: string;
-		description: string;
-		href: string;
-	};
-</script>
-
 <script lang="ts">
 	import { Article, FooterNav, MainFooter } from '$lib/components';
 	import { Main } from '$lib/layouts';
 	import { getContext } from 'svelte';
-	import type { PostsData } from './Post.svelte';
-	import type { AuthorData } from './Author.svelte';
+	import type { PostsData, AuthorData} from '$routes/blog/content';
 	import { BLOG_TITLE_SUFFIX } from '$routes/titles';
 	import { DEFAULT_HOST } from '$lib/utils/metadata';
