+++ src/markdoc/layouts/Post.svelte
@@ -1,23 +1,8 @@
-<script context="module" lang="ts">
-    export type PostsData = {
-        title: string;
-        description: string;
-        date: Date;
-        cover: string;
-        timeToRead: number;
-        author: string;
-        category: string;
-        href: string;
-        featured?: boolean;
-    };
-</script>
-
 <script lang="ts">
     import { Article, FooterNav, MainFooter, Newsletter } from '$lib/components';
     import { Main } from '$lib/layouts';
     import { getContext } from 'svelte';
-    import type { AuthorData } from './Author.svelte';
-    import type { CategoryData } from './Category.svelte';
+    import type { CategoryData, AuthorData, PostsData } from '$routes/blog/content';
     import { BLOG_TITLE_SUFFIX } from '$routes/titles';
     import { DEFAULT_HOST } from '$lib/utils/metadata';
